use new simplifier API for integers_from.
This was wrong with the new API, and as a result it was inheriting
simplifiers from IntStrategy, so doing The Wrong Thing. This broke
Fraction because it would simplify it to Fraction(_, 0).

This also adds testing of Decimal and Fractions so as to catch this
error more clearly rather than as part of other tests.